+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 0.1 (18 Feb 2026)
+- Added multi-language support (no GUI for changing the language currently)
+- Added encryption tab
+- Added startup sequence
+
 ## 0.0 (14 Feb 2026)
 - Started project